plugin samples
