+++ samples/plugins/README.md
@@ -4,5 +4,7 @@
 
 Pack
 
+go build -o out -buildmode=plugin ./plugin/*.go
+
 Start dubbogo provider in directory in [server](../dubbogo/http/server)
 Start dubbogo proxy using the [config](config/api_config.yaml)
